Reorganize code in preparation for v5.1
