+++ ddht/v5/messages.py
@@ -1,20 +1,10 @@
-from collections import UserDict
-from typing import TYPE_CHECKING, Type
-
-from eth_utils import int_to_big_endian
-import rlp
 from rlp.sedes import Binary, CountableList, big_endian_int, binary, boolean
 
+from ddht.base_message import BaseMessage
 from ddht.constants import IP_V4_SIZE, IP_V6_SIZE
 from ddht.enr import ENR
+from ddht.message_registry import MessageTypeRegistry
 from ddht.v5.constants import TOPIC_HASH_SIZE
-
-# https://github.com/python/mypy/issues/5264#issuecomment-399407428
-if TYPE_CHECKING:
-    MessageTypeRegistryBaseType = UserDict[int, Type["BaseMessage"]]
-else:
-    MessageTypeRegistryBaseType = UserDict
-
 
 #
 # Custom sedes objects
@@ -33,52 +23,20 @@ class IPAddressSedes(Binary):  # type: ignore
 ip_address_sedes = IPAddressSedes()
 
 
-class MessageTypeRegistry(MessageTypeRegistryBaseType):
-    def register(self, message_data_class: Type["BaseMessage"]) -> Type["BaseMessage"]:
-        """Class Decorator to register BaseMessage classes."""
-        message_type = message_data_class.message_type
-        if message_type is None:
-            raise ValueError("Message type must be defined")
-
-        if message_type in self:
-            raise ValueError(f"Message with type {message_type} is already defined")
-
-        if not self:
-            expected_message_type = 1
-        else:
-            expected_message_type = max(self.keys()) + 1
-
-        if not message_type == expected_message_type:
-            raise ValueError(
-                f"Expected message type {expected_message_type}, but got {message_type}"
-            )
-
-        self[message_type] = message_data_class
-
-        return message_data_class
-
-
-default_message_type_registry = MessageTypeRegistry()
+v5_registry = MessageTypeRegistry()
 
 
 #
 # Message types
 #
-class BaseMessage(rlp.Serializable):  # type: ignore
-    message_type: int
-
-    def to_bytes(self) -> bytes:
-        return b"".join((int_to_big_endian(self.message_type), rlp.encode(self)))
-
-
-@default_message_type_registry.register
+@v5_registry.register
 class PingMessage(BaseMessage):
     message_type = 1
 
     fields = (("request_id", big_endian_int), ("enr_seq", big_endian_int))
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class PongMessage(BaseMessage):
     message_type = 2
 
@@ -90,14 +48,14 @@ class PongMessage(BaseMessage):
     )
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class FindNodeMessage(BaseMessage):
     message_type = 3
 
     fields = (("request_id", big_endian_int), ("distance", big_endian_int))
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class NodesMessage(BaseMessage):
     message_type = 4
 
@@ -108,14 +66,14 @@ class NodesMessage(BaseMessage):
     )
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class ReqTicketMessage(BaseMessage):
     message_type = 5
 
     fields = (("request_id", big_endian_int), ("topic", topic_sedes))
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class TicketMessage(BaseMessage):
     message_type = 6
 
@@ -126,21 +84,21 @@ class TicketMessage(BaseMessage):
     )
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class RegTopicMessage(BaseMessage):
     message_type = 7
 
     fields = (("request_id", big_endian_int), ("ticket", binary), ("node_record", ENR))
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class RegConfirmationMessage(BaseMessage):
     message_type = 8
 
     fields = (("request_id", big_endian_int), ("registered", boolean))
 
 
-@default_message_type_registry.register
+@v5_registry.register
 class TopicQueryMessage(BaseMessage):
     message_type = 9
